refactor: 스크린들 screens 디렉토리에 이동, 메인 메뉴에 대해서 모듈화
