+++ lib/screens/mainpage/Deployment.dart
@@ -1,8 +1,7 @@
+import 'package:easyback/screens/main_menu.dart';
 import 'package:flutter/material.dart';
 import '../baepopage/baepo.dart';
 import '../baepopage/sangtae.dart';
-import 'ApiPage.dart';
-import '../loginpage.dart';
 
 class Deployment extends StatefulWidget {
   const Deployment({Key? key});
@@ -33,65 +32,7 @@ class _DeploymentState extends State<Deployment> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.black,
-      appBar: PreferredSize(
-        preferredSize: Size.fromHeight(100),
-        child: AppBar(
-          leading: Container(),
-          title: Row(
-            children: [
-              Padding(
-                padding: EdgeInsets.only(top: 20.0, bottom: 8.0, left: 8.0),
-                child: IconButton(
-                  icon: Image.asset(
-                    'asset/image/great.png',
-                    width: 60,
-                    height: 60,
-                  ),
-                  onPressed: () {
-                    _handleImagePressed(context);
-                  },
-                ),
-              ),
-              Expanded(
-                flex: 4,
-                child: TextButton(
-                  onPressed: () {
-                    _handleAPIButton(context);
-                  },
-                  child: Text(
-                    'API',
-                    style: TextStyle(color: Colors.white),
-                  ),
-                ),
-              ),
-              Expanded(
-                flex: 16,
-                child: Align(
-                  alignment: Alignment.centerLeft,
-                  child: TextButton(
-                    onPressed: () {
-                      _handleDeploymentButton(context);
-                    },
-                    child: Text(
-                      'Deployment',
-                      style: TextStyle(color: Colors.white),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          backgroundColor: Colors.black,
-          actions: [
-            IconButton(
-              icon: Icon(Icons.person, color: Colors.white),
-              onPressed: () {
-                _handleLoginPageButton(context);
-              },
-            ),
-          ],
-        ),
-      ),
+      appBar: MainMenu(),
       body: Column(
         crossAxisAlignment: CrossAxisAlignment.start, // 왼쪽 정렬
         children: [
@@ -259,34 +200,10 @@ class _DeploymentState extends State<Deployment> {
     );
   }
 
-  void _handleAPIButton(BuildContext context) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => ApiPage()),
-    );
-  }
-
-  void _handleDeploymentButton(BuildContext context) {
-    setState(() {
-      centerText = null; // 배포 가이드 버튼 클릭 시 텍스트 초기화
-    });
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => Deployment()),
-    );
-  }
-
   void _handlesangtaeButton(BuildContext context) {
     Navigator.push(
       context,
       MaterialPageRoute(builder: (context) => Sangtae()),
-    );
-  }
-
-  void _handleLoginPageButton(BuildContext context) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => loginpage()),
     );
   }
 
@@ -295,9 +212,5 @@ class _DeploymentState extends State<Deployment> {
       context,
       MaterialPageRoute(builder: (context) =>Baepo()),
     );
-  }
-
-  void _handleImagePressed(BuildContext context) {
-    Navigator.popUntil(context, ModalRoute.withName('/'));
   }
 }
